Remove sd-decrypt vm
Minimal changes introduced to move sd-decrypt logic in to the sd-svs vm.
Removed config-dvm as it appeared deprecated and no longer in use.
Removed accept-sd-xfer-extracted mime handlers as logic is now on same vm that process decryptions.
