--- dom0/sd-decrypt-files.top
@@ -1,6 +1,0 @@
-# -*- coding: utf-8 -*-
-# vim: set syntax=yaml ts=2 sw=2 sts=2 et :
-
-base:
-  sd-decrypt:
-    - sd-decrypt-files
