+++ dom0/sd-svs.sls
@@ -22,3 +22,9 @@ sd-svs-dom0-qubes.OpenInVM:
   file.prepend:
     - name: /etc/qubes-rpc/policy/qubes.OpenInVM
     - text: "sd-svs $dispvm:sd-svs-disp allow\n"
+
+# Allow sd-svs to access gpg keys on sd-gpg
+sd-svs-dom0-qubes.qubesGpg:
+  file.prepend:
+    - name: /etc/qubes-rpc/policy/qubes.Gpg
+    - text: "sd-svs sd-gpg allow\n"
