--- sd-decrypt/config-dvm
@@ -1,14 +1,0 @@
-#!/usr/bin/bash
-
-# this gets run in the DVM template
-
-gpg --import /home/user/QubesIncoming/dom0/sd-journalist.sec
-
-sudo cp /home/user/QubesIncoming/dom0/decrypt/decrypt-sd-submission.desktop /usr/share/applications/
-sudo cp /home/user/QubesIncoming/dom0/decrypt/application-x-sd-xfer.xml /usr/share/mime/packages/
-sudo cp /home/user/QubesIncoming/dom0/decrypt/decrypt-sd-submission /usr/bin
-
-sudo update-mime-database /usr/share/mime
-sudo update-desktop-database /usr/share/applications
-
-touch /home/user/.qubes-dispvm-customized
