--- sd-svs/application-x-sd-xfer-extracted.xml
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<mime-info xmlns="http://www.freedesktop.org/standards/shared-mime-info">
-  <mime-type type="application/x-sd-xfer-extracted">
-    <comment>Archive for transfering files from disp vm to sd-svs vm for viewing and storing.</comment>
-    <glob pattern="*.sd-xfer-extracted"/>
-  </mime-type>
-</mime-info>
